レイアウトを改善
