+++ .tool-versions
@@ -1,0 +1,1 @@
+pnpm 10
