+++ pages/welcome.vue
@@ -2,15 +2,16 @@
 import { signIn } from '~/lib/auth-client';
 
 definePageMeta({
+  layout: false,
   middleware: undefined
 })
 </script>
 
 <template>
-  <div class="p-3 h-screen grid place-items-center">
+  <main class="p-3 h-screen grid place-items-center">
     <ClientOnly>
       <UiBlurReveal :delay="0.5" :duration="0.75">
-        <h1 class="text-4xl text-center my-4 lg:text-6xl font-bold">OK Mailへようこそ！</h1>
+        <h1 class="text-4xl text-center my-4 sm:text-6xl font-bold">OK Mailへようこそ！</h1>
         <p class="text-center text-lg">
           楽しいメール管理体験を、無料で始めましょう。
         </p>
@@ -26,5 +27,5 @@ definePageMeta({
         </div>
       </UiBlurReveal>
     </ClientOnly>
-  </div>
+  </main>
 </template>
